Refactor basic algorithm tests into separate modules
